+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Optional callback `widgetNotFound` in `requestWidgetUpdate`
+- Exposed `getWidgetInfo` function that returns a list of widgets added on the home screen
+
 ## [0.1.1] - 2023-02-19
 
 ### Fixed